Move logic into a mixin so it can be effectively themed
